리뷰 :
1. 메인페이지 - 리뷰 섹션 : best review 섹션
   - 현재 bomiora.kr에서도 best review 의 카드 누르면 해당 카드의 리뷰가 보여지는게 아니라 전체 리뷰 페이지로 이동 = 왜냐면 리뷰가 가짜 (정부현이 만들어낸거라서 내용이 없음)
   - 비대면 상품은 (4개의별 + 좋았던점/아쉬운점/꿀팁)을 써야하지만 정부현이 만들어낸것은 제목+내용+4개의 별 밖에 없음.
   - so, 가짜로 만들어낸 best review list 페이지 생성
   - 메인의 best reivew card를 누르면 best review list page로 이동함.
   - best review list page : 인플루언서 닉네임 + 제목 + 내용 + 4개의 별 (서포터리뷰 레이아웃 똑같이 사용)

2. 일반 제품 상세 - 서포터 리뷰 탭 생성
   - 고명진이 없앤거 정부현의 쉐이크때문에 다시 생성

3. 일반 제품 리뷰 수정으로 인한 일반상세페이지 리뷰 탭 리스트 수정
   - 양한열이 고명진의 기획을 읽지 않고 제품 상세 리뷰 페이지를 제작하였기때문에 이 페이지 다시 기획함으로 인해 레이아웃 수정 개발 작업 진행

4. 리뷰 작성/수정 페이지 리뷰 정책 푸터 생성

5. 비대면 제품 상세
   - 한의약품 배지를 가격 -% ( )-> 여기다가 위치 이동

6. 기타
  - 비대면리뷰 감량 프로그레스 바 기준 수정
  - 알럿 대신 공통_토스트오버레이로 validation 체크 표시
  - 제품 정보 섹션 하드코딩 제거 -> 내용 db에서 가져오도록 수저

+++ lib/data/models/review/review_model.dart
@@ -12,11 +12,11 @@ class ReviewModel {
   final DateTime? isTime;
   final int? isConfirm;
   
-  // 평점 (각 5점 만점, 정수 1~5 — DB `is_score1~4` INT)
-  final int isScore1; // 효과
-  final int isScore2; // 가성비
-  final int isScore3; // 맛/향
-  final int isScore4; // 편리함
+  // 평점 (각 5점 만점, 0.1 단위 — DB `is_score1~4` DECIMAL)
+  final double isScore1; // 효과
+  final double isScore2; // 가성비
+  final double isScore3; // 맛/향
+  final double isScore4; // 편리함
   /// 처방·일반 공통 상품 만족도 0.1~5, 0.1 단위 (API `totalIsScore` / DB `total_is_score`)
   final double? totalIsScore;
   final double? averageScore; // 평균 평점
@@ -53,10 +53,10 @@ class ReviewModel {
   bool get isSupporterReview => isRvkind == 'supporter';
   bool get isGeneralReview => isRvkind == 'general';
   bool get isSatisfied => isRecommend == 'y';
-  int get score1 => isScore1;
-  int get score2 => isScore2;
-  int get score3 => isScore3;
-  int get score4 => isScore4;
+  double get score1 => isScore1;
+  double get score2 => isScore2;
+  double get score3 => isScore3;
+  double get score4 => isScore4;
   
   ReviewModel({
     this.isId,
@@ -277,10 +277,10 @@ class ReviewModel {
       isName: NodeValueParser.asString(normalized['isName']),
       isTime: NodeValueParser.asDateTime(normalized['isTime']),
       isConfirm: NodeValueParser.asInt(normalized['isConfirm']),
-      isScore1: NodeValueParser.asInt(normalized['isScore1']) ?? 0,
-      isScore2: NodeValueParser.asInt(normalized['isScore2']) ?? 0,
-      isScore3: NodeValueParser.asInt(normalized['isScore3']) ?? 0,
-      isScore4: NodeValueParser.asInt(normalized['isScore4']) ?? 0,
+      isScore1: NodeValueParser.asDouble(normalized['isScore1']) ?? 0,
+      isScore2: NodeValueParser.asDouble(normalized['isScore2']) ?? 0,
+      isScore3: NodeValueParser.asDouble(normalized['isScore3']) ?? 0,
+      isScore4: NodeValueParser.asDouble(normalized['isScore4']) ?? 0,
       totalIsScore: NodeValueParser.asDouble(normalized['totalIsScore'] ?? normalized['total_is_score']),
       averageScore: NodeValueParser.asDouble(normalized['averageScore']),
       isRvkind: NodeValueParser.asString(normalized['isRvkind']) ?? 'general',
